testability improvements
