+++ lib/router.auth.js
@@ -10,6 +10,7 @@ const AuditLog = require("./thinx/audit"); let alog = new AuditLog();
 const Database = require("./thinx/database.js");
 
 const sha256 = require("sha256");
+const testEnvi = process.env.ENVIRONMENT === "test" ? require("../spec/_envi.json") : null;
 
 module.exports = function (app) {
 
@@ -26,6 +27,56 @@ module.exports = function (app) {
     let userlib = require("nano")(db_uri).use(prefix + "managed_users"); // lgtm [js/unused-local-variable]
 
     const InfluxConnector = require('./thinx/influx');
+
+    function ensureTestUser(username, callback) {
+        if (process.env.ENVIRONMENT !== "test" || !testEnvi || username !== testEnvi.dynamic.username) {
+            callback(false);
+            return;
+        }
+
+        const dynamicUser = testEnvi.dynamic;
+        const dynamicUserBody = {
+            first_name: dynamicUser.first_name,
+            last_name: dynamicUser.last_name,
+            email: dynamicUser.email,
+            username: dynamicUser.username
+        };
+        const passwordHash = sha256(prefix + dynamicUser.username);
+
+        const ensurePassword = () => {
+            user.apply_update(dynamicUser.owner, "password", passwordHash, () => {
+                callback(true);
+            });
+        };
+
+        user.profile(dynamicUser.owner, (success) => {
+            if (success) {
+                ensurePassword();
+                return;
+            }
+
+            user.create(dynamicUserBody, true, {}, (_res, createSuccess, response) => {
+                if (createSuccess && typeof (response) === "string" && response.length === 64) {
+                    user.set_password({
+                        password: dynamicUser.username,
+                        rpassword: dynamicUser.username,
+                        activation: response
+                    }, () => {
+                        callback(true);
+                    });
+                    return;
+                }
+
+                if (response === "email_already_exists" || response === "username_already_exists") {
+                    ensurePassword();
+                    return;
+                }
+
+                console.log(`⚠️ [warning] Failed to prepare dynamic test user on login: ${response}`);
+                callback(false);
+            });
+        });
+    }
 
     /*
      * Authentication
@@ -273,8 +324,13 @@ module.exports = function (app) {
         user.validate(username, (db_body) => {
 
             if ((typeof (db_body.rows) === "undefined") || (db_body.rows.length == 0)) {
-                console.log(`[OID:0] [LOGIN_INVALID] with username ${username}`);
-                return Util.failureResponse(res, 403, "invalid_credentials");
+                return ensureTestUser(username, (prepared) => {
+                    if (prepared) {
+                        return loginAction(req, res);
+                    }
+                    console.log(`[OID:0] [LOGIN_INVALID] with username ${username}`);
+                    return Util.failureResponse(res, 403, "invalid_credentials");
+                });
             }
 
             let user_data = db_body.rows[0].value;
